converted to floats

+++ include/superpixels/RegionSplitter.h
@@ -7,15 +7,15 @@ class RegionSplitter
     public:
         RegionSplitter(const SuperpixelParams & params);
 
-        void run(const std::vector<std::vector<double>> & centers,
+        void run(const std::vector<std::vector<float>> & centers,
                     const std::vector<std::vector<cv::Point>> & superpixels,
-                    std::vector<std::vector<Eigen::Vector2d>> & superpixel_projections,
-                    std::vector<Eigen::Matrix3d> & superpixel_rotations,
+                    std::vector<std::vector<Eigen::Vector2f>> & superpixel_projections,
+                    std::vector<Eigen::Matrix3f> & superpixel_rotations,
                     const cv::Mat & depth_image);
 
     private:
 
-        bool polarSort(const Eigen::Vector2d & a, const Eigen::Vector2d & b);
+        bool polarSort(const Eigen::Vector2f & a, const Eigen::Vector2f & b);
 
         SuperpixelParams params_;    
         rclcpp::Logger logger_ = rclcpp::get_logger("RegionSplitter");
