+++ src/ConvexHullifier.cpp
@@ -10,32 +10,32 @@ void ConvexHullifier::setParams(const SuperpixelParams & params)
     params_ = params;
 }
 
-Eigen::Vector3d ConvexHullifier::projectPointOntoPlane(const Eigen::Vector3d & regionPt)
+Eigen::Vector3f ConvexHullifier::projectPointOntoPlane(const Eigen::Vector3f & regionPt)
 {
-    Eigen::Vector3d planeCenter = Eigen::Vector3d(0, 0, 0); // center is the origin
-    Eigen::Vector3d planeNormal = Eigen::Vector3d(0, 0, 1);
-    Eigen::Vector3d diff = regionPt - planeCenter; 
+    Eigen::Vector3f planeCenter = Eigen::Vector3f(0, 0, 0); // center is the origin
+    Eigen::Vector3f planeNormal = Eigen::Vector3f(0, 0, 1);
+    Eigen::Vector3f diff = regionPt - planeCenter; 
     float dist = diff.dot(planeNormal);
-    Eigen::Vector3d projPt = regionPt - dist * planeNormal;
+    Eigen::Vector3f projPt = regionPt - dist * planeNormal;
 
     return projPt;
 }
 
-void ConvexHullifier::run(std::vector<std::vector<double>> & centers,
+void ConvexHullifier::run(std::vector<std::vector<float>> & centers,
                             std::vector<int> & center_counts,
                             std::vector<std::vector<cv::Point>> & superpixels,
-                            std::vector<std::vector<Eigen::Vector2d>> & superpixel_projections,
-                            std::vector<std::vector<Eigen::Vector2d>> & superpixel_convex_hulls,
-                            std::vector<Eigen::Matrix3d> & egocan_to_region_rotations,
+                            std::vector<std::vector<Eigen::Vector2f>> & superpixel_projections,
+                            std::vector<std::vector<Eigen::Vector2f>> & superpixel_convex_hulls,
+                            std::vector<Eigen::Matrix3f> & egocan_to_region_rotations,
                             const cv::Mat & depth_img)
 {
     // RCLCPP_INFO_STREAM(logger_, "   [ConvexHullifier::run]");
 
-    // std::vector<std::vector<double>> centers_lc = centers;
+    // std::vector<std::vector<float>> centers_lc = centers;
     // std::vector<std::vector<cv::Point>> superpixels_lc = superpixels;
-    superpixel_projections = std::vector<std::vector<Eigen::Vector2d>>(centers.size());
-    superpixel_convex_hulls = std::vector<std::vector<Eigen::Vector2d>>(centers.size());
-    egocan_to_region_rotations = std::vector<Eigen::Matrix3d>(centers.size());
+    superpixel_projections = std::vector<std::vector<Eigen::Vector2f>>(centers.size());
+    superpixel_convex_hulls = std::vector<std::vector<Eigen::Vector2f>>(centers.size());
+    egocan_to_region_rotations = std::vector<Eigen::Matrix3f>(centers.size());
 
     // Build convex hulls for each superpixel
     for (int i = 0; i < (int) centers.size(); i++)
@@ -67,30 +67,30 @@ void ConvexHullifier::run(std::vector<std::vector<double>> & centers,
         cv::Vec3f centerEgocanPt;
         pixelToEgocanFrame(centerEgocanPt, center_pixel, center_depth, params_.k_c_, params_.h_);
 
-        Eigen::Vector3d center(centerEgocanPt.val[0], centerEgocanPt.val[1], centerEgocanPt.val[2]);
-        Eigen::Vector3d normal(centers[i][3], centers[i][4], centers[i][5]);
+        Eigen::Vector3f center(centerEgocanPt.val[0], centerEgocanPt.val[1], centerEgocanPt.val[2]);
+        Eigen::Vector3f normal(centers[i][3], centers[i][4], centers[i][5]);
 
         // RCLCPP_INFO_STREAM(logger_, "       center (egocan frame): " << center.transpose());
         // RCLCPP_INFO_STREAM(logger_, "       normal (egocan frame): " << normal.transpose());
 
-        Eigen::Vector3d arbitraryVec(1, 0, 0);
+        Eigen::Vector3f arbitraryVec(1, 0, 0);
 
         // check here to make sure arbitraryVec is not parallel to normal
         if (std::abs(arbitraryVec.dot(normal)) > 0.99)
         {
-            arbitraryVec = Eigen::Vector3d(0, 1, 0);
+            arbitraryVec = Eigen::Vector3f(0, 1, 0);
         }
 
-        Eigen::Vector3d e0 = normal.cross(arbitraryVec);
+        Eigen::Vector3f e0 = normal.cross(arbitraryVec);
         e0.normalize();
-        Eigen::Vector3d e1 = normal.cross(e0);
+        Eigen::Vector3f e1 = normal.cross(e0);
         e1.normalize();
 
         // RCLCPP_INFO_STREAM(logger_, "       e0: " << e0.transpose());
         // RCLCPP_INFO_STREAM(logger_, "       e1: " << e1.transpose());
 
         // Egocan to region rotation matrix
-        Eigen::Matrix3d egocanToRegionRotMat;
+        Eigen::Matrix3f egocanToRegionRotMat;
         // almost positive this is correct
         egocanToRegionRotMat.row(0) = e0;
         egocanToRegionRotMat.row(1) = e1;
@@ -103,36 +103,36 @@ void ConvexHullifier::run(std::vector<std::vector<double>> & centers,
         // RCLCPP_INFO_STREAM(logger_, "           " << egocanToRegionRotMat.row(1));
         // RCLCPP_INFO_STREAM(logger_, "           " << egocanToRegionRotMat.row(2));
 
-        // Eigen::Quaterniond regionQuat(egocanToRegionRotMat);
+        // Eigen::Quaternionf regionQuat(egocanToRegionRotMat);
 
         // RCLCPP_INFO_STREAM(logger_, "               center: " << center.transpose());
         // RCLCPP_INFO_STREAM(logger_, "               regionQuat: " << regionQuat.x() << ", " << regionQuat.y() << ", " << regionQuat.z() << ", " << regionQuat.w());
 
-        // Eigen::VectorXd region_pose_world_frame = transformHelperPoseStamped(center, regionQuat, egocanFrameToWorldFrame);
+        // Eigen::VectorXf region_pose_world_frame = transformHelperPoseStamped(center, regionQuat, egocanFrameToWorldFrame);
 
         // // get rotation matrix
-        // Eigen::Matrix3d rotMat = calculateRotationMatrix(region_pose_world_frame[3], region_pose_world_frame[4], region_pose_world_frame[5]);
+        // Eigen::Matrix3f rotMat = calculateRotationMatrix(region_pose_world_frame[3], region_pose_world_frame[4], region_pose_world_frame[5]);
 
-        // Eigen::Matrix4d worldToRegionTransform = Eigen::Matrix4d::Identity();
+        // Eigen::Matrix4f worldToRegionTransform = Eigen::Matrix4f::Identity();
         // worldToRegionTransform.block<3, 3>(0, 0) = rotMat;
         // worldToRegionTransform.block<3, 1>(0, 3) = region_pose_world_frame.head(3); 
 
-        // Eigen::Matrix4d egocanToWorldTransform = Eigen::Matrix4d::Identity();
+        // Eigen::Matrix4f egocanToWorldTransform = Eigen::Matrix4f::Identity();
         // // quaternion to rotation matrix
-        // Eigen::Quaterniond egocanToWorldQuat(egocanFrameToWorldFrame.transform.rotation.w, 
+        // Eigen::Quaternionf egocanToWorldQuat(egocanFrameToWorldFrame.transform.rotation.w, 
         //                                         egocanFrameToWorldFrame.transform.rotation.x, 
         //                                         egocanFrameToWorldFrame.transform.rotation.y, 
         //                                         egocanFrameToWorldFrame.transform.rotation.z);
-        // Eigen::Matrix3d egocanToWorldRotMat = egocanToWorldQuat.toRotationMatrix();
+        // Eigen::Matrix3f egocanToWorldRotMat = egocanToWorldQuat.toRotationMatrix();
         // egocanToWorldTransform.block<3, 3>(0, 0) = egocanToWorldRotMat;
-        // Eigen::Vector3d egocanToWorldTrans(egocanFrameToWorldFrame.transform.translation.x,
+        // Eigen::Vector3f egocanToWorldTrans(egocanFrameToWorldFrame.transform.translation.x,
         //                                     egocanFrameToWorldFrame.transform.translation.y,
         //                                     egocanFrameToWorldFrame.transform.translation.z);
         // egocanToWorldTransform.block<3, 1>(0, 3) = egocanToWorldTrans;
 
-        // Eigen::Matrix4d egocanToRegionTransform = egocanToWorldTransform * worldToRegionTransform;
+        // Eigen::Matrix4f egocanToRegionTransform = egocanToWorldTransform * worldToRegionTransform;
 
-        Eigen::Matrix4d egocanToRegionTransform;
+        Eigen::Matrix4f egocanToRegionTransform;
         egocanToRegionTransform << egocanToRegionRotMat, -egocanToRegionRotMat * center,
                                     0, 0, 0, 1;
 
@@ -169,15 +169,15 @@ void ConvexHullifier::run(std::vector<std::vector<double>> & centers,
 
             // RCLCPP_INFO_STREAM(logger_, "               (egocan frame): " << egocanPt.val[0] << ", " << egocanPt.val[1] << ", " << egocanPt.val[2]);
 
-            Eigen::Vector4d egocanPtHomog(egocanPt.val[0], egocanPt.val[1], egocanPt.val[2], 1.0);
-            Eigen::Vector4d regionPtHomog = egocanToRegionTransform * egocanPtHomog;
+            Eigen::Vector4f egocanPtHomog(egocanPt.val[0], egocanPt.val[1], egocanPt.val[2], 1.0);
+            Eigen::Vector4f regionPtHomog = egocanToRegionTransform * egocanPtHomog;
 
-            Eigen::Vector3d regionPt = regionPtHomog.head(3);
+            Eigen::Vector3f regionPt = regionPtHomog.head(3);
 
             // RCLCPP_INFO_STREAM(logger_, "               (region frame): " << regionPt.transpose());
 
             // project points onto plane
-            Eigen::Vector3d projPt = projectPointOntoPlane(regionPt);
+            Eigen::Vector3f projPt = projectPointOntoPlane(regionPt);
 
             // RCLCPP_INFO_STREAM(logger_, "               projected (region frame): " << projPt.transpose());
 
@@ -196,7 +196,7 @@ void ConvexHullifier::run(std::vector<std::vector<double>> & centers,
         // {
             // RCLCPP_INFO_STREAM(logger_, "                   " << superpixel_convex_hulls[i][j][0] << ", " << superpixel_convex_hulls[i][j][1]);
 
-            // if (superpixel_convex_hulls[i][j] == Eigen::Vector2d(0.0, 0.0)) // if a hull point is too close to origin
+            // if (superpixel_convex_hulls[i][j] == Eigen::Vector2f(0.0, 0.0)) // if a hull point is too close to origin
             // {
             //     RCLCPP_WARN_STREAM(logger_, "   [ConvexHullifier::run]");
             //     RCLCPP_WARN_STREAM(logger_, "       center " << i << ":");
@@ -218,15 +218,15 @@ void ConvexHullifier::run(std::vector<std::vector<double>> & centers,
 
             //         RCLCPP_WARN_STREAM(logger_, "               (egocan frame): " << egocanPt.val[0] << ", " << egocanPt.val[1] << ", " << egocanPt.val[2]);
 
-            //         Eigen::Vector4d egocanPtHomog(egocanPt.val[0], egocanPt.val[1], egocanPt.val[2], 1.0);
-            //         Eigen::Vector4d regionPtHomog = egocanToRegionTransform * egocanPtHomog;
+            //         Eigen::Vector4f egocanPtHomog(egocanPt.val[0], egocanPt.val[1], egocanPt.val[2], 1.0);
+            //         Eigen::Vector4f regionPtHomog = egocanToRegionTransform * egocanPtHomog;
 
-            //         Eigen::Vector3d regionPt = regionPtHomog.head(3);
+            //         Eigen::Vector3f regionPt = regionPtHomog.head(3);
 
             //         RCLCPP_WARN_STREAM(logger_, "               (region frame): " << regionPt.transpose());
 
             //         // project points onto plane
-            //         Eigen::Vector3d projPt = projectPointOntoPlane(regionPt);
+            //         Eigen::Vector3f projPt = projectPointOntoPlane(regionPt);
 
             //         RCLCPP_WARN_STREAM(logger_, "               projected (region frame): " << projPt.transpose());
             //     }
@@ -239,13 +239,13 @@ void ConvexHullifier::run(std::vector<std::vector<double>> & centers,
     return;
 }
 
-void ConvexHullifier::convexHull(const std::vector<Eigen::Vector2d> & superpixel_projections, 
-                                    std::vector<Eigen::Vector2d> & convex_hull)
+void ConvexHullifier::convexHull(const std::vector<Eigen::Vector2f> & superpixel_projections, 
+                                    std::vector<Eigen::Vector2f> & convex_hull)
 {
     grahamScan(superpixel_projections, convex_hull);
 }
 
-bool ConvexHullifier::polarSort(const Eigen::Vector2d & a, const Eigen::Vector2d & b, const Eigen::Vector2d & lowest)
+bool ConvexHullifier::polarSort(const Eigen::Vector2f & a, const Eigen::Vector2f & b, const Eigen::Vector2f & lowest)
 {
     if (a == lowest)
     {
@@ -256,8 +256,8 @@ bool ConvexHullifier::polarSort(const Eigen::Vector2d & a, const Eigen::Vector2d
         return false;
     }
 
-    double angle_a = std::atan2(a[1] - lowest[1], a[0] - lowest[0]);
-    double angle_b = std::atan2(b[1] - lowest[1], b[0] - lowest[0]);
+    float angle_a = std::atan2(a[1] - lowest[1], a[0] - lowest[0]);
+    float angle_b = std::atan2(b[1] - lowest[1], b[0] - lowest[0]);
 
     if (angle_a < angle_b)
     {
@@ -273,15 +273,15 @@ bool ConvexHullifier::polarSort(const Eigen::Vector2d & a, const Eigen::Vector2d
     }
 }
 
-bool ConvexHullifier::ccw(const Eigen::Vector2d & a, const Eigen::Vector2d & b, const Eigen::Vector2d & c)
+bool ConvexHullifier::ccw(const Eigen::Vector2f & a, const Eigen::Vector2f & b, const Eigen::Vector2f & c)
 {
-    Eigen::Vector2d a_to_b = b - a;
-    Eigen::Vector2d a_to_c = c - a;
+    Eigen::Vector2f a_to_b = b - a;
+    Eigen::Vector2f a_to_c = c - a;
     return (a_to_b[0] * a_to_c[1] - a_to_b[1] * a_to_c[0]) > 0;
 }
 
-void ConvexHullifier::grahamScan(const std::vector<Eigen::Vector2d> & superpixel_projections, 
-                                    std::vector<Eigen::Vector2d> & convex_hull)
+void ConvexHullifier::grahamScan(const std::vector<Eigen::Vector2f> & superpixel_projections, 
+                                    std::vector<Eigen::Vector2f> & convex_hull)
 {
     // RCLCPP_INFO_STREAM(logger_, "           [ConvexHullifier::grahamScan]");
 
@@ -294,11 +294,11 @@ void ConvexHullifier::grahamScan(const std::vector<Eigen::Vector2d> & superpixel
 
     // RCLCPP_INFO_STREAM(logger_, "               finding lowest y-coordinate ... ");
     // Find the point with the lowest y-coordinate
-    Eigen::Vector2d lowest = superpixel_projections[0];
+    Eigen::Vector2f lowest = superpixel_projections[0];
     // RCLCPP_INFO_STREAM(logger_, "                   initial lowest: " << lowest[0] << ", " << lowest[1]);
     for (int i = 1; i < (int) superpixel_projections.size(); i++)
     {
-        Eigen::Vector2d current = superpixel_projections[i];
+        Eigen::Vector2f current = superpixel_projections[i];
         // RCLCPP_INFO_STREAM(logger_, "                   point " << i << ": " << current[0] << ", " << current[1]);
         if (current[1] < lowest[1])
         {
@@ -320,7 +320,7 @@ void ConvexHullifier::grahamScan(const std::vector<Eigen::Vector2d> & superpixel
     // RCLCPP_INFO_STREAM(logger_, "           sorting ... ");
 
     // Sort the points by polar angle with respect to the lowest point
-    std::vector<Eigen::Vector2d> sorted = superpixel_projections;
+    std::vector<Eigen::Vector2f> sorted = superpixel_projections;
     std::sort(sorted.begin(), sorted.end(), std::bind(&ConvexHullifier::polarSort, 
                                                         this, 
                                                         std::placeholders::_1, 
@@ -330,7 +330,7 @@ void ConvexHullifier::grahamScan(const std::vector<Eigen::Vector2d> & superpixel
     // RCLCPP_INFO_STREAM(logger_, "           sorted.");
 
     // Perform the Graham scan
-    std::vector<Eigen::Vector2d> hull;
+    std::vector<Eigen::Vector2f> hull;
     hull.push_back(sorted[0]);
     // RCLCPP_INFO_STREAM(logger_, "           added point 0: " << sorted[0][0] << ", " << sorted[0][1]);
     hull.push_back(sorted[1]);
